requirements.txt
bot-studio==1.4.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
filelock==3.9.0
fonttools==4.39.0
huggingface-hub==0.13.1
idna==3.4
install==1.3.5
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
pyenchant==3.2.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post1
threadpoolctl==3.1.0
tokenizers==0.13.2
tqdm==4.65.0
transformers==4.26.1
typing_extensions==4.5.0
urllib3==1.26.15
youtube-comment-scraper-python==1.0.0
